pyproject.toml
[project]
name = "personal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "jsonpatch>=1.33",
    "pydantic>=2.11.4",
]
